feat: Pass page diagnostics to topical authority flow
This change modifies the `topicalAuthorityFlow` to accept `page_diagnostics` as an input parameter. This data will be used by the Pillar Discovery Agent to inform its analysis of strategic pillars, potentially improving the accuracy and relevance of pillar identification.

The API route and client-side component have been updated to pass this new parameter.